+++ app/src/main/java/com/example/mealmanagement/model/UserInfo.java
@@ -2,7 +2,7 @@ package com.example.mealmanagement.model;
 
 public class UserInfo {
 
-    long id;
+    int id;
     String name;
     String mail;
     String password;
@@ -20,11 +20,11 @@ public class UserInfo {
         this.approve = 0;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
